+++ source/Point.h
@@ -69,8 +69,8 @@ public:
 	double LengthSquared() const;
 	Point Unit() const;
 
-	//Linearly interpolate between the caller to the target by float distance.
-	Point Lerp(const Point &target, const float &distance) const;
+	//Linearly interpolate between the caller to the target by distance.
+	Point Lerp(const Point &target, const double &distance) const;
 
 	double Distance(const Point &point) const;
 	double DistanceSquared(const Point &point) const;
